fix/refactoring, * change progect structure * add folders Components, libs * срфchange files destination
